debian/patches: Ensure that identify works with old goodix driver
Newer fprintd supports duplicate-detection, this implies sending to the
device an empty prints gallery, expecting the device to give us a
matched print against its storage.

In case the device has no storage we're doing this anyway, but goodix
doesn't handle this properly, so let's just ignore the case if no prints
are passed.

Also, once returned the value we might initialize a type multiple times,
and let's avoid this not to have a dead-lock after the first action.

LP: #1966911